requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
ollama>=0.5.3
pytest>=7.0.0
httpx[socks]>=0.24.0
black>=24.1.0
ruff>=0.4.0
pre-commit>=3.6.0
anthropic>=0.15.0